+++ eslint.config.js
@@ -1,4 +1,6 @@
 import { tsConfig } from '@jill64/eslint-config-ts'
 
 /** @type {import('@jill64/eslint-config-ts').FlatConfig[]} */
-export default tsConfig()
+export default tsConfig({
+  tsConfigPath: './eslint.tsconfig.json'
+})
