+++ tsconfig.json
@@ -5,5 +5,6 @@
     "declaration": true,
     "module": "NodeNext",
     "moduleResolution": "NodeNext"
-  }
+  },
+  "include": ["src"]
 }
